# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Document API",
        "title": "Geolocation",
        "contact": {
            "name": "API Support",
            "url": "https://simpplify.com.br/contact",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/tractor-unit/check-plate/{plate}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Verifica se a placa existe ou já está cadastrada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cavalo"
                ],
                "summary": "Verificar Placa.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Placa do Veículo",
                        "name": "plate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado da verificação da placa",
                        "schema": {}
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tractor-unit/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria uma unidade tratora.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cavalo"
                ],
                "summary": "Criar uma Unidade Tratora.",
                "parameters": [
                    {
                        "description": "Requisição de Unidade Tratora",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tractor_unit.CreateTractorUnitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações da Unidade Tratora",
                        "schema": {
                            "$ref": "#/definitions/tractor_unit.TractorUnitResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tractor-unit/delete/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui uma unidade tratora.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cavalo"
                ],
                "summary": "Excluir uma Unidade Tratora.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Unidade Tratora",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tractor-unit/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recupera as informações da unidade tratora.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cavalo"
                ],
                "summary": "Obter Unidade Tratora.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Unidade Tratora",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações da Unidade Tratora",
                        "schema": {
                            "$ref": "#/definitions/tractor_unit.TractorUnitResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tractor-unit/list/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recupera as informações do reboque a partir do ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carroceria"
                ],
                "summary": "Obter Carroceria por ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Carroceria",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Carroceria",
                        "schema": {
                            "$ref": "#/definitions/trailer.TrailerResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tractor-unit/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza os dados de uma unidade tratora.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cavalo"
                ],
                "summary": "Atualizar uma Unidade Tratora.",
                "parameters": [
                    {
                        "description": "Requisição de Unidade Tratora",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tractor_unit.UpdateTractorUnitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações da Unidade Tratora",
                        "schema": {
                            "$ref": "#/definitions/tractor_unit.TractorUnitResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "tractor_unit.CreateTractorUnitRequest": {
            "type": "object",
            "properties": {
                "axles": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "can_couple": {
                    "type": "boolean"
                },
                "capacity": {
                    "type": "string"
                },
                "chassis": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "engine_power": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "length": {
                    "type": "number"
                },
                "license_plate": {
                    "type": "string"
                },
                "manufacture_year": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "renavan": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "unit_type": {
                    "type": "string",
                    "enum": [
                        "stump",
                        "truck",
                        "tractor_unit"
                    ]
                },
                "user_id": {
                    "type": "integer"
                },
                "width": {
                    "type": "number"
                }
            }
        },
        "tractor_unit.TractorUnitResponse": {
            "type": "object",
            "properties": {
                "axles": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "can_couple": {
                    "type": "boolean"
                },
                "capacity": {
                    "type": "string"
                },
                "chassis": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "engine_power": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "number"
                },
                "license_plate": {
                    "type": "string"
                },
                "manufacture_year": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "renavan": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "unit_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "width": {
                    "type": "number"
                }
            }
        },
        "tractor_unit.UpdateTractorUnitRequest": {
            "type": "object",
            "properties": {
                "axles": {
                    "type": "integer"
                },
                "brand": {
                    "type": "string"
                },
                "capacity": {
                    "type": "string"
                },
                "chassis": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "engine_power": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "number"
                },
                "license_plate": {
                    "type": "string"
                },
                "manufacture_year": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "renavan": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "unit_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "width": {
                    "type": "number"
                }
            }
        },
...
        "trailer.TrailerResponse": {
            "type": "object",
            "properties": {
                "axles": {
                    "type": "integer"
                },
                "body_type": {
                    "type": "string"
                },
                "chassis": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "length": {
                    "type": "number"
                },
                "license_plate": {
                    "type": "string"
                },
                "load_capacity": {
                    "type": "number"
                },
                "renavan": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "width": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}